webadmin: Add quotaId and quotaName to Cloner
Updating the hosted engine VM via webadmin was failing due to the
values for quotaId and quotaName not being cloned in the Cloner, which
left the values set to null in the clone. Those values are not
editable for the HE VM, leading to a failure during validation.

This patch adds these values to the Cloner in order to  prevent the
validation failure.

Change-Id: I45af78e71d1a624967c9427f3672746a3e0855b9
Bug-Url: https://bugzilla.redhat.com/1410431
Signed-off-by: Phillip Bailey <phbailey@redhat.com>