aaa: Session expired message shown on SSO login page after session expiration
When the user leaves the login page open
for a period of time the session expires
and the associated SsoSession is cleaned
up. Entering user credentials and attempting
to login shows "Session expired please try
again" message.

The SsoSession object should be saved in
SsoContext with an identification token as
key which is also available on login page.
When user tried to login if the session has
expired the SsoSession can be retrieved from
the SsoContext using the session id token.

Change-Id: Idc33455751ea25e8ac77b5549e503d15b7a06960
Bug-Url: https://bugzilla.redhat.com/1411416
Signed-off-by: Ravi Nori <rnori@redhat.com>