core : Storage domains not showing in search on a DC named with an underscore
When data center name has an underscore in them
the search sql for storage domains for the data
center is executed with escaped data center name.

This causes issue when the escaped name is used
in the left hand side of the like clause. When
the name is used on the left hand side of the
like clause it does not have to be escaped.

Change-Id: I3049eb2641d7a86b0b857aa9a46376d794d14812
Bug-Url: https://bugzilla.redhat.com/1409849
Signed-off-by: Ravi Nori <rnori@redhat.com>