core: RunVmCommand - Emulated Machine conforms the CCV
RunVmCommand selects 'default' Emulated Machine corresponding
to the Custom Compatibility Version, if set.

The selection is based on best-match heuristics to match the platform
(recent default cluster emulated machine vs. vdc_options.ClusterEmulatedMachine).

The issue was originaly identified for preview of snapshot with memory
originating in previous cluster version (before cluster upgrade).
In that flow, the CCV is set for the previewed snapshot to
retain VM configuration, including the emulated machine.

Change-Id: I029441010cfd4c3b47fbe687c96748aced565bae
Bug-Url: https://bugzilla.redhat.com/1364456
Signed-off-by: Marek Libra <mlibra@redhat.com>
Signed-off-by: Arik Hadas <ahadas@redhat.com>