webadmin: Improve message about unmanaged network
Previously, the message about operations on unmanaged or unsynchronized
networks could be interpreted misleading, since it does not contain the
information what network the message is about.

The name of the unmanaged or unsynchronized network is included in the
message, to point out what network the message refers to.

Change-Id: Ia2f5fc3f78d438ba31bfefa4d97ee5789fb72427
Bug-Url: https://bugzilla.redhat.com/1167675
Signed-off-by: Dominik Holler <dholler@redhat.com>
(cherry picked from commit f2fb9c92093abd265a1c93265450462a3b7452ba)