core: monitoring vm devices using the domain xml
The monitoring of vm devices can now be done according to libvirt's
domain xml rather than the dictionary we used to get from VDSM.
However, in order to minimize the changes in vm devices monitoring,
we convert the devices in the domain xml into VDSM-like maps that
contain the relevant data, but this is done on the engine side.

Note that the DomainXML config value still defaults to false, so
we still call FullList by default (this is going to be changed
separately, once the monitoring of host devices is added).

Change-Id: I04e03e1137567ff8a14c2b07b1e8f02d1a832b8c
Signed-off-by: Arik Hadas <ahadas@redhat.com>