core: Don't use anySetOf(Class)
Mockito 1.x's anySet() returned a raw Set, which would cause
unchecked assignment warning when it was used in a generic context.
Mockito 2.x fixed this to return a Set<T>, and uses Java 8's
capabilities to infer the generic type. As a result, anySetOf(Class)
was deprecated, and will be removed in the near future.

This patch removes all usages of anySetOf(Class) and replaces them with
the recommended anySet().

Change-Id: I4fe8615f84a11035a8564ef4b6767bedf2c8cbad
Signed-off-by: Allon Mureinik <amureini@redhat.com>