core: Don't use anyCollectionOf(Class)
Mockito 1.x's anyCollection() returned a raw Collection, which would
cause unchecked assignment warning when it was used in a generic
context.
Mockito 2.x fixed this to return a Collection<T>, and uses Java 8's
capabilities to infer the generic type. As a result,
anyCollectionOf(Class) was deprecated, and will be removed in the near
future.

This patch removes all usages of anyCollectionOf(Class) and replaces
them with the recommended anyCollection().

Change-Id: I4506932c45937dd9e15172f289a8bcea097f6fc0
Signed-off-by: Allon Mureinik <amureini@redhat.com>