core: Don't use anyListOf(Class)
Mockito 1.x's anyList() returned a raw List, which would cause
unchecked assignment warning when it was used in a generic context.
Mockito 2.x fixed this to return a List<T>, and uses Java 8's
capabilities to infer the generic type. As a result, anyListOf(Class)
was deprecated, and will be removed in the near future.

This patch removes all usages of anyListOf(Class) and replaces them with
the recommended anyList().

Note that this patch only applies to the backend. The fronted is
compiled with a Java 7 language level, and thus cannot rely on type
inference.

Change-Id: I7c90929e93a2e339b673cf54f4ef57063bdf2ddf
Signed-off-by: Allon Mureinik <amureini@redhat.com>