core: Don't use anyMapOf(Class, Class)
Mockito 1.x's anyMap() returned a raw Map, which would cause
unchecked assignment warning when it was used in a generic context.
Mockito 2.x fixed this to return a Map<T, S>, and uses Java 8's
capabilities to infer the generic type. As a result,
anyMapOf(Class, Class) was deprecated, and will be removed in the near
future.

This patch removes all usages of anyMapOf(Class, Class) and replaces
them with the recommended anyMap().

Change-Id: Iec2032f642430b4cde15a3d8b29e00cd6becd047
Signed-off-by: Allon Mureinik <amureini@redhat.com>