core: External VMs addition fix
Problem: Bean validation of VmBase#maxMemorySizeMb was failing in
AddVmFromScratchCommand since it value was 0 (out of valid range). The
AddVmFromScratchCommand is called in order to add external VM to engine.
As result reporting of external VMs was broken.

Fix: Value of #maxMemorySizeMb is property copied from VDSM description
of external VM in AddUnmanagedVmsCommand#convertVm.

If 'maxMemSize' property is mising in the VDSM description then
maxMemory is set to value of memory.

Change-Id: Ie43865b3a15537bd244e3217b926955f814cfe0c
Bug-Url: https://bugzilla.redhat.com/1411739
Signed-off-by: Jakub Niedermertl <jniederm@redhat.com>