core: use default mac pool for clusters without DC
There might exist clusters, which are not linked to any data center.
For those first update statement does not work, leaving possible null
values in cluster.mac_pool_id column, and because of that adding of
constraint will fail.

We have to select some mac pool to be used for such clusters, and
having no information which should be used, we can use only default
mac pool.

Change-Id: If33a212d6c8f730bae8ebe46bc1d6359e95c0495
Bug-Url: https://bugzilla.redhat.com/1410189
Signed-off-by: Martin Mucha <mmucha@redhat.com>