engine: Add timeout to communication with network provider
Previously there was no explicit timeout for communication to external
network provider. This way there was a long timeout applied to the
communication. Since some objects are locked during the communication,
some objects was locked for a long period of time if the external
network provided did not respond.

An explicit timeout for the communication is added, to avoid locking for
a long period of time.

The library com.woorea.openstack is used for the communication with
external network providers. The version of this library is increased to
3.1.2, which introduces the possibility to add the timeout for
communication.

Change-Id: Ic2eb1380273c811bf659ac067e6ebb65cbf31b99
Bug-Url: https://bugzilla.redhat.com/1380356
Signed-off-by: Dominik Holler <dholler@redhat.com>
(cherry picked from commit 5f7c65771a412e5952d693de0bddffea1a59244d)