restapi: Consolidate TestHelper params matchers
With the changes in ArgumentMatcher that were delivered in the upgrade
to Mockito 2.x, the ActionParameterEquals and QueryParameterEquals
matchers became almost completely identical, except for their generic
specification.

This patch applies a generic T parameter to the matchers and unifies
them to a single ParametersEquals matcher which saves code bloating and
is easier to maintain.

Change-Id: If5716f42c9309250a1856c04d3bc77673f79eeea
Signed-off-by: Allon Mureinik <amureini@redhat.com>