packaging: setup: Offer to fix db misconfigured items
Motivation
A setup may fail with misconfigured db properties, but it fails-fast,
and that may send the user into a loop untill all the db values are
configured correctly while instead we can report all the missing
items at once and offer to fix them

Change
gather all db config errors and report at once.
If the user approves and the db is localhost - the setup will try to fix
the configuration file and restart postgres (notice, won't work for
dev-env).
If the db is a remote db, the user will get a warning with the
misconfigured items.

Change-Id: I99d5a661044a4077589326b8f215fa2152952304
Bug-Url: https://bugzilla.redhat.com/1388433
Signed-off-by: Roy Golan <rgolan@redhat.com>