restapi: Implement 'StorageDomainDisks' and 'AttachedStorageDomainDisks'
Currently we use the same 'Disks' service definitions fo the following
disk collections:

  /disks
  /storagedomain/{sd:id}/disks
  /datacenters/{dc:id}/storagedomains/{sd:id}/disks

This is wrong, because most operations on disk are then available in
multiple places, and it prevents addition of new operations that only
make sense for disks that aren't yet part of the system. A good example
of this is the operation to register an unregistered disk. That only
makes sense for disks that appear in /storagedomains/{sd:id}/disks. To
improve the situation this patch implements the new 'StorageDomainDisks'
and 'AttachedStorageDomainDisk' that have been introduced in the
specification of the API by the following patch:

  Add 'StorageDomainDisks' and 'AttachedStorageDomainDisks'
  https://gerrit.ovirt.org/69634

After this, each collection will have its own service:

  /disks
  DisksService

  /storagedomain/{sd:id}/disks
  StorageDomainDisksService

  /datacenters/{dc:id}/storagedomains/{sd:id}/disks
  AttachedStorageDomainDisksService

These services will initially preserve most of the duplicated
operations, for backwards compatibility. Those duplicated operations
will be marked as deprecated, and removed in a future version of the
API.

Change-Id: I1021d75609a7b07e6ebea279dd4e91ed37973725
Related-To: https://bugzilla.redhat.com/1366960
Related-To: https://gerrit.ovirt.org/69634
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>