webadmin: default storage domain on VM import
When importing a VM from an export domain, each disk's destination
storage domain is selected by default to be the first one that shows in
the list.
This patch changes this default selection to be the storage domain with
the most free space of all the domains in the list, making a scenario
where the default selection is a storage domain that has not enough
space less possible.

Change-Id: Iea988fea5de32b44353c526d2bfe119d3e13ab9c
Bug-Url: https://bugzilla.redhat.com/1279407
Signed-off-by: Idan Shaby <ishaby@redhat.com>