core: Check storage domain before executing commands with VM leases
Before creating or updating a VM with a VM lease, validate that the
selected storage domain for the lease exists, active and is a part of the
storage pool the VM belongs to, if those terms are not met the operation
will fail.
In case of importing a VM with a VM lease, if the domain does not meet the
criteria the VM will be imported without a VM lease and an audit log
warning will be issued.

Bug-Url: https://bugzilla.redhat.com/1413961
Change-Id: Id448f6e86d635ffe1743ed8e542d631d8ea2fe5a
Signed-off-by: Tal Nisan <tnisan@redhat.com>