core: improve the robusness of vm devices monitoring
We should not add a device with no general type or device type.
Before this patch, we generated a log but then a NPE occurred
and the whole devices monitoring cycle was terminated - none of
the devices was updated.
This patch changes this so we will actually skip the invalid
device and continue to process the valid devices.

Change-Id: I74349fa53b9a23661924a66d4c20056e41cb3d08
Signed-off-by: Arik Hadas <ahadas@redhat.com>