core: StorageJobCallback - prevent NPE when there is no executing host
When there is no executing host (may happen when hosts go no responsive
during the command execution for example), the StorageJobCallback may fail
with NPE when attempting to check the host status.

This patch prevents it, if there is no executing host the job status can
be considered as failed.

Change-Id: I394c6b6a88bee9ef58d25b4e7bd3e880685b9100
Bug-Url: https://bugzilla.redhat.com/1414675
Signed-off-by: Liron Aravot <laravot@redhat.com>