webadmin: add ui logic to WAD-EnableDiscard relationship
Previously, the user could tick the wipe after delete and enable discard
checkboxes as he wished, and only after the ok button was clicked and
the request was sent to the backend, an error would have raised in case
of an illegal configuration and the user didn't have a chance to fix it
since the popup was closed.
This patch adds the same backend logic of enable discard and its
relationship with wipe after delete to the ui in order to save the user
from disappointing after finding out about the wrong configuration too
late.

For more information about enable discard and its relationship with wipe
after delete, please refer to the feature page:
http://www.ovirt.org/develop/release-management/features/storage/pass-discard-from-guest-to-underlying-storage/

Change-Id: I912ae85e5d6dfa7bd5aad17d41d0e051e78f457e
Bug-Url: https://bugzilla.redhat.com/1414126
Signed-off-by: Idan Shaby <ishaby@redhat.com>