restapi: Fork to execute the metamodel tool
Currently the metamodel tool, which is used to generate part of the API
code, is executed within the same Java virtual machine that runs Maven.
In some recent versions of Maven this generates a situation where Maven
is not able to stop the threads created by the CDI container (Weld SE)
created by the tool. To avoid that issue this patch changes the build
process so that the metamodel tool runs inside its own Java virtual
machine.

Change-Id: I87f5ced1e0a6adf9ff5f5652eb6ca0d08b9081ce
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>