engine: normalize bonding options without miimon
The bonding options are normalized by placing the option "mode"
in the first place and the option "miimon" in the second place
in the options string.

Previously, if the option "miimon" was missing, a space was added
the first place of the options string instead of the present option.

If the option "miimon" is missing, the option "mode" is still
positioned in the first place.

Change-Id: Ie3417fe2dc5bffb1b98c6a664da1c0be87fc54be
Related-To: https://bugzilla.redhat.com/1412563
Signed-off-by: Dominik Holler <dholler@redhat.com>