core: StorageJobCallback - jobs fencing
When a job status can't be detected using entity polling the
StorageJobCallback will attempt to fence it in case it hasn't
started yet (when a job starts it's persisted in the command
parameters).

This patch changes the logic to attempt to fence the job even if it has
started, in case the job status can't be determined using entity polling
performing the fencing logic can confirm us that the operation has
failed on the next entity polling round.
Additionally, if the job has started and is stuck on a storage operation
we can still attempt to fence it using the entity. As the job should
modify the entity only after acquiring a cluster lock on it.

Change-Id: Ifdbab1c6e898b7791e6b05b62684f036c23cc3d4
Bug-Url: https://bugzilla.redhat.com/1415502
Signed-off-by: Liron Aravot <laravot@redhat.com>