engine: Upgrade JUnit dependency to 4.12
Upgrade the JUnit dependency to 4.12, that last available 4.x version.

Besides the general notion of staying current and moving ahead with the
times, this version offers several specific important improvements
that affect our codebase, without any changes made:
* Performance improvement on assertArrayEquals (see PR #918)
* TimeoutExceptions now contain stacktraces from stuck threads (see PR
  #742, PR #986)
* Improved @Theory failure messages (see PR #607)

Additional improvements require changes in the codebase in order to be
consumed, and would be introduced in following patches.

Change-Id: I4c67d754b5230b533e4ae56b818972e2903bf50f
Signed-off-by: Allon Mureinik <amureini@redhat.com>