core: Fix storage domain creation rollback on failure
When adding a new storage domain fails with an exception, the compensation
mechanism should kick in and rollback all newly created entities, among
them the storage domain, however since patch e7b5093 introduced, a command
execution was added as a part of the flow that adds a default disk profile.
This command is using the same context of the parent command resulting in
clearing the old compensation context and upon failure only the disk
profile entity creation is rolled back and the storage domain entity
remains in the DB resulting in the UI inability to remove the connection
and an unusable storage domain entity existing in the database.
This patch fixes it by cloning the context to the add disk profile command
and detaching it from the parent context.

Change-Id: I3d42062313792f680d0f4e6e51ee5837f34a08ec
Bug-Url: https://bugzilla.redhat.com/1404601
Signed-off-by: Tal Nisan <tnisan@redhat.com>