webadmin: initialize TabIndexes with 1 by default
As defined at MDN [1], tabindex should be set to a positive
value, as value '0' means that the element should be placed
after elements with a positive tabindex.

For now, this change fixes tab order in 'New Disk' dialog.
I.e. when 'Size' is focused, pressing 'tab' should focus on
'Alias' (and not on search bar, as it did previously).

[1] https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/tabindex

Change-Id: I91dbedee38d893302f1006c9ce1e58d146ce7cce
Bug-Url: https://bugzilla.redhat.com/??????
Signed-off-by: Daniel Erez <derez@redhat.com>