core: Update QCOW volumes in 4.0 DC to the default QCOW version.
After upgrade of oVirt 4.1 if there were any added disks in a 4.0 DC
those disks will not be initiailzed with the default QCOW version since
the update could be done only if VDSM reported the QCOW version of the
new volume which was only introduced in 4.1 DCs.

Therefore any added QCOW volumes in DC 4.0 should be initialized with
the default QCOW version which is 0.10.

Also added an upgrade script that updates all the QCOW volumes
that were added to a 4.0 DC after the oVirt upgrade, to the default
QCOW version.

Change-Id: Ibddd5a8bcac7e13aa98cc2c0bb9f8a48d4ac40f0
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>