core: extract common code for building vm-info
Code that is common to both the generation of VM info map/dict
and libvirt's Domain XML is extracted from VmInfoBuilderImpl
to VmInfoBuildUtils. In general, the builders should not have
code that is not directly related to the conversion to the
target representation and can be extracted to a utility method.

Change-Id: I72c9df7f03c42ef4405a9159fa7e84ca76cec64c
Signed-off-by: Arik Hadas <ahadas@redhat.com>