core: Change max vCPU count computation
Maximum vCPU count for a VM is currently determined by a defined
absolute maximum limit and the number of cores and threads. However, on
x86 the simple formula may cause overflow of an APIC ID limit due to the
bit-based layout of the APIC ID value and then QEMU refuses to run the
VM.

With this patch we make sure that the value always fits into the APIC
limits. See the bug report for explanation. In the future, when x2APIC
will be used, we can remove the restriction.

Change-Id: I9cced4317eebfd302230152e96c6c32dcf0fb13f
Bug-Url: https://bugzilla.redhat.com/1406243
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>