core: always compute boot order on run vm
On regular run-VM we used to rely on the boot order that is set on each
device, while on run-once we computed the boot order according to the
given boot sequence (if it is different than the default boot sequence
that is defined for the VM, otherwise it behaved as regular run).

With this patch we always compute the boot order, not only on
run-once flow (on regular run it is done according to the VM's default
boot sequence). That way, we don't rely on the boot order that is set
on each device, so it enables us to remove this field and get rid of
the duplicated data whose management tends to complicate addition,
update and removal of bootable devices.

Change-Id: I20b6ad6c2002a857b073dd81a32be064415fbbcc
Signed-off-by: Arik Hadas <ahadas@redhat.com>