core: do not treat video device as bootable device
Recent change in VmInfoBuilderImpl added video device to the collection named
managedDevices. This collection is only used to set the boot order on bootable
devices. Since video device is not a bootable device, it should be added to
that collection.

The name of the collection is confusing, thus this patch renames it so it will
better reflect its intended use.

Change-Id: If2108bd0b25015c8bd9a3ba9344dc5954bba90c2
Signed-off-by: Arik Hadas <ahadas@redhat.com>