core: prepare updateBootOrder for templates
VmDeviceUtils#updateBootOrder could have work only with VMs - it assumed
that the ID it receives is an ID of a VM. This patch is a step toward
breaking this assumption. It is needed because once we stop persisting
VmDevice#bootOrder, we will need to compute the boot order before
exporting a template.

Change-Id: Id82afd0c2270eb0e8c9292700e69f10c1be92fd2
Signed-off-by: Arik Hadas <ahadas@redhat.com>