core: remove redundant legacy check in create vm
Generally, we do not have validations in VDS command. That is because the bll
command validates that the operation can be executed and the locks it takes
ensures that this 'valid' state is preserved.

However, we have these annoying checks in create vm. Not only that these
checks are unexpected, but they also interact with the database - and this
has a performance penalty.

One of these checks that the status of the VM is different than ImageLocked.
In RunVm we validate that the VM is not in this status and then exclusively
lock the VM. So this check is relevant only if there is a flow in which the
VM is switched to ImageLocked without being locked by (in-memory) engine
lock. But now that we lock the VM before switching it to ImageLocked in
AddVmTemplate (see If4822f), there is no such flow so this check is
redundant and therefore removed.

Change-Id: I98d9ff00ec182b742f5a04c2866b7852f74c3e92
Signed-off-by: Arik Hadas <ahadas@redhat.com>