webadmin: update actions availability of VMs list with "exclusive" lock
Updating webAdmin actions availability of VMs in VmListModel to treat a VM
in an "exclusive" memory lock and in "Down" status as if it
is in an "ImageLocked" DB status and therefore block relevant actions for it.

This is implemented by adding a new entity for overriding the getStatus()
method of the VM entity, used only by VmListModel.

This fix the problem that till now VMs in an "ImageLocked" status were checked
and relevant actions were blocked for them in Webadmin as required, but VMs in
an "exclusive" memory lock status were treated as if they are in
"Down"/"Up"/Maintenance" etc status so evethough they were in the middle of
a process like importing, relevant actions were not blocked for them as required.

Change-Id: I51776fba35828589604ca44374a7574e91b4a20c
Bug-Url: https://bugzilla.redhat.com/1364137
Signed-off-by: Sharon Gratch <sgratch@redhat.com>