core: send generated domxml to vdsm
With this patch the engine sends Domain XML that is generated by the
engine to VDSM. The XML is added to the VM map/dictionary, so it
would simply be ignored on VDSM versions that do not support it.
New config value is introduced to enable this. For now, its default
value is false but this is expected to change once the implementation
will be more mature.

Change-Id: I4197d4e96645179e3defdcc573cd104defd0225f
Signed-off-by: Arik Hadas <ahadas@redhat.com>