core: RemoveStoragePool - allowed domain statuses
This patch prevents removing a DC while there's a domain which
isn't on Maintenance status.

Currently we allow to remove a DC when there's a domain in Inactive status,
while it still may be accessed by some of the hosts.
This patch changes that and allows to remove a DC only when its domains are on
Maintenance status - an Inactive domains must be deactivated prior to removing the DC.

Change-Id: I784ad9758de867b5092f06761ba9f0880fd6436d
Bug-Url: https://bugzilla.redhat.com/1418020
Signed-off-by: Liron Aravot <laravot@redhat.com>