vdsbroker: no failover on detachSD
The detach operation requires access to the domain as it modifies
its metadata - which will cause us to fail in case the domain
isn't available.
If the domain is unavailable, the detach operation fails and a failover
(assignment of the SPM role to a different host to perform retry) occurs.
As failover should help us rarely when attempting to perform detach,
this patch disables it on that scenario. The SPM can be moved to a
different host by the user if there's an actual issue related to the
host.

Change-Id: Ia6586df4ca6d9302b43bedfeec4fa8070b05412e
Bug-Url: https://bugzilla.redhat.com/1394687
Signed-off-by: Liron Aravot <laravot@redhat.com>