core: treat template's actual disk size as editable field
This field is not being set anyway (it is generated from the template's disks)
so there is no problem in setting it as editable just for the reflection-based
validation to pass on update template.
Proper fix, however, will be to remove the actual disks size from the template.

Change-Id: I216ca19c1a060e0931390d681da8217da316a00a
Bug-Url: https://bugzilla.redhat.com/1417597
Signed-off-by: Arik Hadas <ahadas@redhat.com>
(cherry picked from commit 689eeeb016f84b874308ca7df4e21f050778e28d)