core: VmTemplate.actualDiskSize property not cached
The problem with #actualDiskSize field was that it needed to be marked
as @EditableVmTemplateField to allow UpdateVmTemplateCommand to be
called with new template both with and without `diskImageMap` property
set. This annotation was required even-though no disk related changes
were actually performed by the UpdateVmTemplateCommand.

Related bugs:
https://bugzilla.redhat.com/1417597
https://bugzilla.redhat.com/1409579

From now on VmTemplate.actualDiskSize is a computed property based on
`diskImageMap` property.

The only code that calls getActualDiskSize() repeatedly is
VmTemplateComparerByDiskSize. Cache of `actualDiskSize` values was
introduced in this class.

Change-Id: Ie2011fa8a63c304c220be7a326c4142d81f7cd34
Signed-off-by: Jakub Niedermertl <jniederm@redhat.com>