engine: Support User's JMX invocatio via jconsole
The engine exposes the InMemoryLockManager mbean to view, release a
specific lock or clear all of the acquired locks.

In order to access that mbean there is a need to add native management
interface support. That used to be supported by jboss-cli, but since [1]
it is no longer supported.

In order invoke the InMEmeoryLockManager operations using jconsole:
1. Run on the ovirt-engine server:
 # $WILDFLY_HOME/bin/jconsole.sh
2. Fill the following details:
 * Remote Process Url: service:jmx:remote://localhost:8707
 * Username: admin@internal
 * Password: *****

The username should be the as the once used to connect with jboss-cli.

[1] https://issues.jboss.org/browse/PRODMGT-425

Change-Id: Ide95988516485c5eeb02ed34e7ec534cf5cf4dd8
Bug-Url: https://bugzilla.redhat.com/1417940
Signed-off-by: Moti Asayag <masayag@redhat.com>