engine: Iterable @DataPoints
With the upgrade to JUnit 4.12, Iterables can be used as @DataPoints
(see PR #658).

This patch removes the boiler plate code needed to convert Iterabls
to arrays, in order to make the code cleaner and easier to read.
Note that where using an array was simpler (e.g., SomeEnum#values()
was used a @DataPoints), the method was left untouched.

Change-Id: I8f2fc6a32ab64a28db6dab9da080e1b47817a1f6
Signed-off-by: Allon Mureinik <amureini@redhat.com>