engine: @Parameters return type
With the upgrade to JUnit 4.12, the return types Iterator<? extends
Object>, Object[] and Object[][] are now supported on methods
annotated with @Parameters.

This patch removes the boiler plate code needed to convert Object[][]
(or even Object[]) parameters to Collection(Object[]), in order to
make the code cleaner and easier to read.

Change-Id: Ic31290cab8feeb9e5d49dc4e40a11d08a65d2dfd
Signed-off-by: Allon Mureinik <amureini@redhat.com>