core: more efficient identification of run-once vms
When a VM goes down we check if it ran in run-once mode in order
to see if we need to remove its unmanaged devices. We used to
do that by checking the number of cd-rom/floppy devices although
since 3.3 we have a dedicated flag for that. So from now on we
use this flag instead.

Change-Id: Ib5f2b69680f5cef02c23f8b5d38df4fae581aa49
Signed-off-by: Arik Hadas <ahadas@redhat.com>