core: Avoid NPE when importing unregistered Template.
When importing an unregistered Template the initialization of this
Template is being done only in the validate method since it includes
several DB calls and OVF parsing.

Since the init function of ImportVmTemplateFromConfigurationCommand
executes updateMaxMemorySize which uses getVmTemplate()
we get an NPE and fail to register an unregistered Template.

The following fix uses the same technique which is used for
updateGraphicsDevices by checking if the vmBase is null, and if so
simply return.

Change-Id: If13d8aa654e72569d02cc18e91f19d4af7445ac7
Bug-Url: https://bugzilla.redhat.com/1419364
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>