core: remove another redundant legacy check in create vm
In RunVm we validate that the VM is not during snapshot (that is, that
no snapshot is locked) while the VM is locked and then we check it
again in the create VDS command. The latter check is redundant since
every time that we lock a snapshot (and specifically the active
snapshot) we do it while the VM is locked.

In addition, removing unused method in SnapshotsManager.

Change-Id: I05d8a28e462ff942370fa21477f4524c246daff2
Signed-off-by: Arik Hadas <ahadas@redhat.com>