he_restore: updating foreign key removing HE hosts
For migration or disaster recovery purposes we let the user filter
out the hosted-engine hosts to redeploy from a backup on a
different hosted-engine env. In order to be sure that all the
hosted-engine runs with a local up-do-date configuration we
have to remove them from the engine DB asking to the user to
redeploy them to be sure that they point to the right hosted-engine
storage domain.
In order to remove them at restore time avoiding foreign key errors,
we need to update all the existing VMs related to these hosts.
Set status=Down, exit_reason=AdminShutdown, exit_status=Normal,
run_on_vds=NULL, migrating_to_vds=NULL for all the VM that were running
on the hosted-engine hosts we are filtering out at restore time.

Change-Id: I85a6fcd18005393e69ed319e7b56709e7adc5b9b
Bug-Url: https://bugzilla.redhat.com/1425890
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>