core: Fix OvfManagerTest
@Mock/@Spy annotations do not guarantee any order between them,
making this test inherently racy - if iconDefaultsProvider is
initialized first, it will be available when manager is initialized.
If manager is initialized first, it's will have a null
iconDefaultsProvider, causing some tests to fail.

Since the OsRepository and OvfVmIconDefaultsProvider mocks are read
only anyway, they should be static and initialized ONCE per test case
instead of before every test in such a racy fashion.
As a nice side effect, this will slightly improve the test's runtime.

Change-Id: I5f547092e7934dd73bf2039e9ff08822b95c7fa9
Signed-off-by: Allon Mureinik <amureini@redhat.com>