webadmin: Fix ClassCastException when removing template's disk
When attempting to remove a template's disk from the Storage
sub-tab the operation fails with a ClassCastException.

Fixed by using a list of DiskModel instead of a StorageDomainModel
since we are selecting a disk and operating on it.

Another ClassCastException occurred on line 175 since
result.getState() returns a TemplateStorageListModel and not
a ConfirmationModel.

Change-Id: I81f8d145d4e0ec79f635b1e84589a53fda91f538
Bug-Url: https://bugzilla.redhat.com/1420265
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>