ui: Added 4.2 to the list of NFS versions
This patch adds NFS version 4.2 to the list of NFS versions in the
Storage Domain creation dialog. This allows user to create Storage
Domain that is always mounted with nfsvers=4.2 option.

In particular, this allows user to guarantee that virt-sparsify will be
working on his Storage Domain, because virt-sparsify for its operation
requires version 4.2 of NFS protocol.

Change-Id: I5af47da488cff6219628c0ec29f84ddb8b211768
Bug-Url: https://bugzilla.redhat.com/1406398
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>