Use the overhead calculator to get commited memory in HostMonitoring
The host monitoring is summing up the memory of all VMs that are
currently running on a host. This patch updates the algorithm there
to use the more precise overhead calculation instead of the overhead
constant.

Since the overhead now depends on some static attributes (platform,
devices, ..) I had to reconstruct the VM object from values we
already have available in monitoring and add cache for values
that were missing.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1362557
Change-Id: I426b8f521feb20bda4143a2b7b4b3e49923315a9
Signed-off-by: Martin Sivak <msivak@redhat.com>