webadmin: NPE on edit pool dialog
The resource allocation of the edit pool dialog has a logic which tries to
select the storage domain of the selected template to be the same as the
storage domain on which some Vm from the pool is.

The logic has an assumption that the selected template's disks are the parent
disks of the respective VM. This assumption is true only if the template is not
versioned. If the template has versions, this assumption is broken.

This causes an issue that the disk allocation part is not properly inited and
consequently fails on validation on NPE.

Fixed by first initing the storage to the values of the template disks and than
try to match them to the VM ones.

Change-Id: Iafc386b7cc4b013293c7b7f627053a438f7df7cc
Bug-Url: https://bugzilla.redhat.com/1421725
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>