core: add VM snapshot ID to older glance template disks
An upgrade script to add a VM snapshot ID to disks of glance image
based template disks. This comes in addition to commit 4ddb7947a4
which fixes the issue for new templates in which disks of templates
made of repository images have no VM snapshot ID which causes an
error when attempting to import them from an export domain.

Change-Id: If17ef8a396bc54d6c5e4f3f4e8cfc54670bcfbd4
Bug-Url: https://bugzilla.redhat.com/1399603
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>