core: GlusterTaskUtils map
Anonymously implementing a HashMap just to place values in it is a bad
practice which both "spams" up the classloader with useless anonymous
classes and produces code that's hard to read.

This patch removes the instance of that practice and replaces it with
an old fashioned HashMap with a series of put.

Change-Id: I43dd72c3ef750672775d01fe84dcc341613df0ac
Signed-off-by: Allon Mureinik <amureini@redhat.com>