core: GlusterSnapshotSyncJob maps
Anonymously implementing a HashMap just to place values in it is a bad
practice which both "spams" up the classloader with useless anonymous
classes and produces code that's hard to read.

This patch removes the instance of that practice and replaces it with
an old fashioned HashMap with a series of put.

Change-Id: I1ad0b4612917a735c5818151c8dcb7badfde3f64
Signed-off-by: Allon Mureinik <amureini@redhat.com>