core: GlusterGeoRepSyncJob maps
Anonymously implementing a HashMap just to place values in it is a bad
practice which both "spams" up the classloader with useless anonymous
classes and produces code that's hard to read.

This patch removes the instance of that practice and replaces it with
an old fashioned HashMap with a series of put.

Change-Id: Ibd7bf17ff4608ea63755402fade892e46fc0d106
Signed-off-by: Allon Mureinik <amureini@redhat.com>