db: memory_and_disk_images_storage_domain_view - EXISTS instead of JOIN
Change Ifbedd49da added indexes and changed the
memory_and_disk_images_storage_domain_view view to use JOIN instead of EXISTS
to determine if the disk is a memory disk/ovf store. On certain datasets
and under load that causes postgres to execute the query slower.

As we yet to find a case in which using EXISTS for those tables causes
to a problem, this patch restores the use of it in the view.

Change-Id: I7daa22d37d67502836f5222fc9acaaa85881379e
Bug-Url: https://bugzilla.redhat.com/1412626
Signed-off-by: Liron Aravot <laravot@redhat.com>