core: Adjust the limit on max VCPUs
The maximum number of VCPUs must be multiplication of sockets * cores.  This is
ensured by the formula computing the maximum number of VCPUs, but the
configuration value may not satisfy that constraint.  So if we are limited by
the configuration value then we must check and adjust that value.

This is mostly theoretical in 4.1+ since the current configuration value (288)
is higher than the APIC limit (255) so it should cause no trouble in the
default 4.1 configuration.

Change-Id: I857e1a85950a863752e02a30066f682fff04a7c7
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>