core: Make GlusterTasksSyncJobTest a unit test
GlusterTasksSyncJobTest creates a GlusterTasksSyncJob instance and a
GlusterTaskUtils that's mocked into it.
It then proceeds to test not just GlusterTasksSyncJob's behavior, but
also the inner workings of the GlusterTaskUtils, thus, it is not really
a UNIT test, and maintaining it may become exceedingly complicated when
these inner workings are changed (e.g., if the dependencies of the
GlusterTaskUtils are injected, as is the intention later on in this
patch series).

This patch replaces the real instance of the GlusterTaskUtils with a
mocked one, thus making GlusterTasksSyncJobTest a true unit test.

Change-Id: I53b6a3ed260e57619bad5020a804bb8a93420e47
Signed-off-by: Allon Mureinik <amureini@redhat.com>