core: Generalize GlusterJob scheduling
Currently, there's an awkward coupling between the GlusterJobsManager
that's aware of all the jobs and defines there scheduling policy.

This patch moves the knowledge of the job's schedule to the job (using
the newly created GlusterJobSchedulingDetails class), and makes the
GlusterJobsManager generic - it just loads all the beans extending
GlusterJob, and schedules all of them.

Change-Id: I712c680769c115f9cc76a63391e89841b8b1943e
Signed-off-by: Allon Mureinik <amureini@redhat.com>