webadmin: Unplugging state of hot mem unplug button
When memory hot unplug action is confirmed the "Hot unplug" button is
disabled and relabeled to "Unplugging ..." for 15 seconds. This informs
user that the action is in progress and prevents him from re-invoking
it. "Hot unplug" button is disabled if corresponding VM is not in state
'UP' or if memory hot unplug is disabled by
ConfigValues#HotUnplugMemorySupported.

* New entity VmDeviceFeEntity introduced to be able to hold frontend
  specific state of VmDevice - #isBeingUnplugged. It inherits from
  VmDevice.
* VmDevicesListModel and related classes use VmDeviceFeEntity instead of
  VmDevice.
* backend part: VmDevice has copy constructor and inheritance friendly
  toString() method.

Change-Id: I78a0eba2dd80a01f2f5f695b4ea1034b2b9e40b4
Bug-Url: https://bugzilla.redhat.com/1228543
Signed-off-by: Jakub Niedermertl <jniederm@redhat.com>