core: move initialization type from vm to the parameters of run-vm
It is not really a property of a VM but a configuration for the run operation
(for example, in the next run we may want to use cloud-init and on the next
run we certainly won't want it because the VM is already initialized). Thus,
this patch moves the initilizationType from the overloaded VM object to the
parameters of Create command.

Change-Id: I8e11c3db2ce9fa8501c6f4e8bbe319f47a6be0b9
Signed-off-by: Arik Hadas <ahadas@redhat.com>