core: Increment vm db generation when importing devices
When importing vm devices the db generation column was
not updated.
The ovf generation process relies on this column to know
when a new ovf needs to be generated. Since the column
wasn't updated the imported devices were not included in
the ovf.

Change-Id: I0fe8c7a350e417db1ce0ba273a7d470a89afbe69
Bug-Url: https://bugzilla.redhat.com/1416846
Signed-off-by: Jenny Tokar <jtokar@redhat.com>
(cherry picked from commit df85e2083eae38d52296a574ec35c4831d42ea85)