he_restore: avoid failing with other VM disks
For migration or disaster recovery purposes we let the user
filter out the hosted-engine storage domain
and the engine VM to redeploy from a backup on a different
hosted-engine env.
Due to another bug the user could create a floating
disk on the hosted-engine storage domain and then
connect it to the engine VM.
In this case we should still remove the hosted-engine
storage domain filtering out other VMs disks as well
without failing.

Change-Id: I2be23b172ea592f9b6be08477fad0f4bf29a30ad
Bug-Url: https://bugzilla.redhat.com/1416459
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>