core, webadmin: HotUnplugMemoryCommand added
Command to perform memory hot unplug. Memory device to hot unplug is
identified by VmDeviceId. If the VM is UP and there is a memory device
of that size, VDSM memory hot unplug is called. If the call succeeds and
it wouldn't end up being negative, memory of VM is decreased.
VM#getMinAllocatedMem is also decreased if necessary.

HotUnplugMemoryCommand doesn't use UpdateVmCommand because using it
would either (based on value of
VmManagementParametersBase#applyChangesLater) immediately apply changes
in next run configuration or discard changes planed in next run
configuration.

The command is intentionally non-transactional. Races between writes to
snapshots and vm_static are not considered a problem.

Change-Id: I8e9d09bd01ef357c203ae32cba66f9c59115aeee
Bug-Url: https://bugzilla.redhat.com/1228543
Signed-off-by: Jakub Niedermertl <jniederm@redhat.com>